🚧 misc: move InputChecker functions to StringExtensions
