+++ app/src/main/java/com/bangkit/befitoutfit/ui/component/BottomSheet.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import com.bangkit.befitoutfit.data.model.Outfit
 import com.bangkit.befitoutfit.data.model.Session
 import com.bangkit.befitoutfit.helper.BottomSheetType
-import com.bangkit.befitoutfit.helper.InputChecker.emailChecker
+import com.bangkit.befitoutfit.helper.StringExtensions.emailChecker
 import com.bangkit.befitoutfit.ui.screen.addOutfit.AddOutfitViewModel
 import com.bangkit.befitoutfit.ui.screen.detailOutfit.DetailOutfitViewModel
 import com.bangkit.befitoutfit.ui.screen.profile.ProfileViewModel
